requirements.txt
transformers>=4.40.0
torch>=2.1.0
pillow>=10.0.0
requests>=2.31.0
